refactor: chat과 notification 훅 구조 정리
