+++ src/apis/notification/mutations.ts
@@ -1,7 +1,10 @@
-import { mutationOptions } from '@tanstack/react-query';
+import { type InfiniteData, mutationOptions, type QueryClient } from '@tanstack/react-query';
+import { decrementInboxUnreadCount, setInboxNotificationReadState } from './cache';
+import { notificationQueryKeys } from './queries';
+import type { InboxNotificationListResponse, InboxNotificationUnreadCountResponse } from './entity';
 import { markInboxNotificationAsRead } from '.';
 
-interface MarkInboxNotificationAsReadVariables {
+export interface MarkInboxNotificationAsReadVariables {
   notificationId: number;
   isRead: boolean;
 }
@@ -11,10 +14,25 @@ export const notificationMutationKeys = {
 };
 
 export const notificationMutations = {
-  markInboxAsRead: () =>
+  markInboxAsRead: (queryClient: QueryClient) =>
     mutationOptions({
       mutationKey: notificationMutationKeys.markInboxAsRead(),
       mutationFn: ({ notificationId }: MarkInboxNotificationAsReadVariables) =>
         markInboxNotificationAsRead(notificationId),
+      onSuccess: (_, { notificationId, isRead }) => {
+        queryClient.setQueryData<InfiniteData<InboxNotificationListResponse>>(
+          notificationQueryKeys.inbox.infinite(),
+          (previousData) => setInboxNotificationReadState(previousData, notificationId)
+        );
+
+        if (isRead) {
+          return;
+        }
+
+        queryClient.setQueryData<InboxNotificationUnreadCountResponse>(
+          notificationQueryKeys.inbox.unreadCount(),
+          (previousData) => decrementInboxUnreadCount(previousData)
+        );
+      },
     }),
 };
