+++ src/components/notification/hooks/useInboxNotificationMutations.ts
@@ -1,0 +1,8 @@
+import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { notificationMutations } from '@/apis/notification/mutations';
+
+export const useMarkInboxNotificationAsReadMutation = () => {
+  const queryClient = useQueryClient();
+
+  return useMutation(notificationMutations.markInboxAsRead(queryClient));
+};
